Criação das rotas de alteração no valor da faixa de consumo, listar tabelas, fazer o calculo do consumo de água e deletar tabela. (finalização do projeto)

+++ src/main/java/com/example/tabela_tarifaria_de_agua/repository/TabelaTarifariaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TabelaTarifariaRepository extends JpaRepository<TabelaTarifaria, Long> {
+public interface TabelaTarifariaRepository extends JpaRepository<TabelaTarifaria, Integer> {
     
     Optional<TabelaTarifaria> findByAtivoTrue();
 }
